requirements.txt
alembic==1.8.1
aniso8601==9.0.1
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.2
Flask-Admin==1.6.0
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
outcome==1.2.0
packaging==21.3
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.1.2
wsproto==1.2.0
WTForms==3.0.1
zipp==3.8.1
